requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.11.3
itsdangerous==0.24
python-nmap==0.5.0-1
wsgiref==0.1.2
